Move fsm.hpp & sha_hash.hpp from include/detail subdir to prevent compilation errors in non edm build envs
