+++ include/iso15118/io/connection_ssl.hpp
@@ -7,7 +7,7 @@
 #include <optional>
 
 #include <iso15118/config.hpp>
-#include <iso15118/detail/io/sha_hash.hpp>
+#include <iso15118/io/sha_hash.hpp>
 #include <iso15118/io/poll_manager.hpp>
 
 namespace iso15118::io {
